+++ MatrixSDKTests/MXAutoDiscoveryTests.m
@@ -343,6 +343,47 @@ static NSString *const kIdendityServerPingPath = @"_matrix/identity/api/v1";
 }
 
 
+// Test that MXWellKnown.JSONDictionary keeps original extended data
+- (void)testAutoDiscoveryWellKnownJSONDictionary
+{
+    NSString *baseURL = @"https://myhs.org";
+    
+    NSDictionary *mockBody = @{
+                               @"m.homeserver": @{
+                                       @"base_url" : baseURL
+                                       },
+                               @"im.vector.riot.e2ee": @{
+                                       @"default": @(NO)
+                                   }
+                               };
+    NSDictionary *hsVersionResponse = @{
+                                        @"versions": @[@"r0.4.0"],
+                                        @"unstable_features": @{
+                                                @"m.lazy_load_members": @(YES)
+                                                }
+                                        };
+    [self stubRequestsContaining:kWellKnowPath withJSONResponse:mockBody statusCode:200];
+    [self stubRequestsContaining:kVersionPath withJSONResponse:hsVersionResponse statusCode:200];
+    
+    [self doFindClientConfig:^(XCTestExpectation *expectation, MXDiscoveredClientConfig * _Nonnull discoveredClientConfig) {
+        
+        XCTAssertEqualObjects(discoveredClientConfig.wellKnown.JSONDictionary, mockBody);
+        
+        // Check parsing
+        BOOL isE2EByDefaultEnabledByHSAdmin = YES;
+        MXWellKnown *wellKnown = discoveredClientConfig.wellKnown;
+        if (wellKnown.JSONDictionary[@"im.vector.riot.e2ee"][@"default"])
+        {
+            MXJSONModelSetBoolean(isE2EByDefaultEnabledByHSAdmin, wellKnown.JSONDictionary[@"im.vector.riot.e2ee"][@"default"]);
+        }
+        XCTAssertFalse(isE2EByDefaultEnabledByHSAdmin);
+        
+        [expectation fulfill];
+    }];
+}
+
+
+
 // Test on matrix.org
 // Only for development. Must be disabled for automatic tests
 //- (void)testAutoDiscoveryWithMatrixDotOrg